requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.4
dnspython==2.1.0
docxcompose==1.3.3
docxtpl==0.12.0
et-xmlfile==1.1.0
idna==3.2
Jinja2==3.0.1
lxml==4.6.3
MarkupSafe==2.0.1
openpyxl==3.0.7
Pillow==8.3.1
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
python-docx==0.8.11
python-magic==0.4.24
requests==2.26.0
six==1.16.0
urllib3==1.26.6
